Move main process logic into GL
It becomes nearly impossible to record the proper daily accrual
entries for partial payments without knowing the state of the GL
for the given period. You would have to track the amount each payment
contributed to the payee, and re-calculate how far in advance to start
"eating the a/r" for any partial payment after the first.

Moved all of the process logic into the GL module itself, so it can
more easily inspect it's state as part of the function.

This should also lend itself to some form of GL validation at runtime.
Knowing that the GL is inserting some entry into itself that will cause
an invalid state as it's happening would be sweet.